+++ packages/markstream-octane/src/index.ts
@@ -100,7 +100,6 @@ export type {
   MermaidBlockEvent,
   MermaidBlockNodeProps,
   PreCodeNodeProps,
-  ShikiCodeBlockProps,
 } from './types/component-props'
 export type { NodeComponentProps } from './types/node-component'
 export * from './utils/languageIcon'